edit search cards + update the end contract logic to update the experience level of freelancers or communities

+++ src/main/java/taskaya/backend/DTO/freelancers/requests/AvrHoursPerWeekUpdateRequestDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @Builder
 public class AvrHoursPerWeekUpdateRequestDTO {
-    Double avrgHoursPerWeek;
+    Integer avrgHoursPerWeek;
 }
