+++ src/main/java/taskaya/backend/entity/freelancer/FreelancerBusiness.java
@@ -23,7 +23,7 @@ public class FreelancerBusiness {
     @Column(name = "completed_jobs", nullable = false)
     private Integer completedJobs=0;
 
-    private Double avgHoursPerWeek;
+    private Integer avgHoursPerWeek;
